Some fixes and warnings

+++ Compare.h
@@ -7,12 +7,12 @@ struct ICompare
 {
 	virtual void compare(const couple_t &file_1, const couple_t &file_2) = 0;
 	virtual ~ICompare() = default;
-	//virtual void check() = 0;
+	virtual std::string name() = 0;
 };
 
 struct text_compare : ICompare
 {
-	text_compare() = default;
+	// text_compare() = default;
 	void compare(const couple_t &file_1, const couple_t &file_2) override
 	{
 		if (file_1.second != file_2.second)
@@ -28,6 +28,8 @@ struct text_compare : ICompare
 		}
 		else std::cout << "files " << file_1.first.filename().string() << " and " << file_2.first.filename().string() << " has equivalent size " << file_1.second/1024 << " Kb" << std::endl;
 	}
+	
+	std::string name(){return "text";}
 };
 
 struct copy_compare : ICompare
@@ -48,7 +50,7 @@ struct copy_compare : ICompare
 					std::cout << file_1.first.filename().string() << " copied from " << file_1.first.parent_path().string() << "\t" << (file_2.second - file_1.second)/1024 << " Kb" 
 					<< std::endl;
 				}
-				else std::cout << "File " << (dir_result/file_1.first.filename()).string() << " already exists" << std::endl;
+				else std::cout << "The file " << (dir_result/file_1.first.filename()).string() << " already exists" << std::endl;
 			}
 			else
 			{
@@ -58,7 +60,7 @@ struct copy_compare : ICompare
 					std::cout << file_2.first.filename().string() << " copied from " << file_2.first.parent_path().string() << "\t" << (file_1.second - file_2.second)/1024 << " Kb" 
 					<< std::endl;
 				}
-				else std::cout << "File " << (dir_result/file_2.first.filename()).string() << " already exists" << std::endl;
+				else std::cout << "The file " << (dir_result/file_2.first.filename()).string() << " already exists" << std::endl;
 			}
 		}
 		else std::cout << "Equivalent files " << file_1.first.filename().string() << " and " << file_2.first.filename().string() 
@@ -66,6 +68,8 @@ struct copy_compare : ICompare
 	}
 	else std::cout << "Result directory invalid. Use parameter -o with argument to set result directory" << std::endl;
 	}
+	
+	std::string name(){return "copy";}
 };
 
 struct repl_compare : ICompare	// replace
@@ -75,37 +79,42 @@ struct repl_compare : ICompare	// replace
 	using namespace std;
 	if (!dir_result.empty())
 	{
-		std::uintmax_t difference;
+		std::intmax_t difference;	// signed
 		fs::path rename_path;
 		fs::path result_path;
-		std::string out_str;	// out result string
+		std::string out_str;		// out result string
 		
 		if (file_1.second != file_2.second)
 		{
 			if (file_2.second < file_1.second) //  + (100 * 1024)
 			{
-				difference = (file_1.second - file_2.second)/1024;
-				out_str = "Renamed second " + file_2.first.string() + " " + to_string(difference) + " Kb diff";
+				difference = ((intmax_t)file_1.second - file_2.second)/1024;
+				out_str = "Renamed " + file_2.first.string() + " " + to_string(difference) + " Kb diff";
 				rename_path = file_2.first;
 				result_path = dir_result/file_2.first.filename();	// operator/ preferred separator
 			}
 			else
 			{
-				difference = (file_2.second - file_1.second)/1024;
-				out_str = "Renamed first " + file_1.first.string() + " " + to_string(difference) + " Kb diff";
+				difference = ((intmax_t)file_2.second - file_1.second)/1024;
+				out_str = "Renamed " + file_1.first.string() + " " + to_string(difference) + " Kb diff";
 				rename_path = file_1.first;
 				result_path = dir_result/file_1.first.filename();
 			}
-					
-			// if ((file_2.second - file_1.second) >= (100 * 1024))
-				// std::swap(file_1, file_2);	// need to avoid the second swap
-				
-			fs::rename(rename_path, result_path);
-			std::cout << out_str << std::endl;
+			
+			if (!fs::exists(result_path))
+			{	
+				fs::rename(rename_path, result_path);
+				std::cout << out_str << std::endl;
+			}
+			else std::cout << "File " << result_path.string() << " already exists" << std::endl;
 		}
+		else std::cout << "Equivalent files " << file_1.first.filename().string() << " and " << file_2.first.filename().string()
+			<< "\t" << file_2.second/1024 << " Kb" << std::endl;
 	}
 	else std::cout << "Result directory invalid. Use parameter -o with argument to set result directory" << std::endl;
 	}
+	
+	std::string name(){return "rename";}
 };
 
 }
